move parseOptions into helper
