+++ src/co2.js
@@ -3,144 +3,7 @@
 import OneByte from "./1byte.js";
 import SustainableWebDesign from "./sustainable-web-design.js";
 
-function parseOptions(options) {
-  // CHeck that it is an object
-  if (typeof options !== "object") {
-    throw new Error("Options must be an object");
-  }
-
-  const adjustments = {};
-
-  if (options?.gridIntensity) {
-    adjustments.gridIntensity = {};
-    const { device, dataCenter, network } = options.gridIntensity;
-    if (device) {
-      if (typeof device === "object" && device.country) {
-        const averageIntensity =
-          require("./data/average-intensities-2021.min.js").default;
-        if (!averageIntensity.data[device.country]) {
-          throw new Error(
-            `"${device.country}" is not a valid country. Please use a valid 3 digit ISO 3166 country code. \nSee https://developers.thegreenwebfoundation.org/co2js/data/ for more information.`
-          );
-        }
-        adjustments.gridIntensity["device"] = {
-          country: device.country,
-          value: parseFloat(averageIntensity.data[device.country]),
-        };
-      } else if (typeof device === "number") {
-        adjustments.gridIntensity["device"] = {
-          value: device,
-        };
-      } else {
-        throw new Error(
-          `The device grid intensity must be a number or an object. You passed in a ${typeof device}.`
-        );
-      }
-    }
-    if (dataCenter) {
-      if (typeof dataCenter === "object" && dataCenter.country) {
-        const averageIntensity =
-          require("./data/average-intensities-2021.min.js").default;
-        if (!averageIntensity.data[dataCenter.country]) {
-          throw new Error(
-            `"${dataCenter.country}" is not a valid country. Please use a valid 3 digit ISO 3166 country code. \nSee https://developers.thegreenwebfoundation.org/co2js/data/ for more information.`
-          );
-        }
-        adjustments.gridIntensity["dataCenter"] = {
-          country: dataCenter.country,
-          value: parseFloat(averageIntensity.data[dataCenter.country]),
-        };
-      } else if (typeof dataCenter === "number") {
-        adjustments.gridIntensity["dataCenter"] = {
-          value: dataCenter,
-        };
-      } else {
-        throw new Error(
-          `The data center grid intensity must be a number or an object. You passed in a ${typeof dataCenter}.`
-        );
-      }
-    }
-    if (network) {
-      if (typeof network === "object" && network.country) {
-        const averageIntensity =
-          require("./data/average-intensities-2021.min.js").default;
-        if (!averageIntensity.data[network.country]) {
-          throw new Error(
-            `"${network.country}" is not a valid country. Please use a valid 3 digit ISO 3166 country code. \nSee https://developers.thegreenwebfoundation.org/co2js/data/ for more information.`
-          );
-        }
-        adjustments.gridIntensity["network"] = {
-          country: network.country,
-          value: parseFloat(averageIntensity.data[network.country]),
-        };
-      } else if (typeof network === "number") {
-        adjustments.gridIntensity["network"] = {
-          value: network,
-        };
-      } else {
-        throw new Error(
-          `The network grid intensity must be a number or an object. You passed in a ${typeof network}.`
-        );
-      }
-    }
-  }
-
-  if (options?.dataReloadRatio) {
-    if (typeof options.dataReloadRatio === "number") {
-      if (options.dataReloadRatio > 0 && options.dataReloadRatio < 1) {
-        adjustments.dataReloadRatio = options.dataReloadRatio;
-      } else {
-        throw new Error(
-          `The dataReloadRatio option must be a number between 0 and 1. You passed in ${options.dataReloadRatio}.`
-        );
-      }
-    } else {
-      throw new Error(
-        `The dataReloadRatio option must be a number. You passed in a ${typeof options.dataReloadRatio}.`
-      );
-    }
-  }
-
-  if (options?.firstVisitPercentage) {
-    if (typeof options.firstVisitPercentage === "number") {
-      if (
-        options.firstVisitPercentage > 0 &&
-        options.firstVisitPercentage < 1
-      ) {
-        adjustments.firstVisitPercentage = options.firstVisitPercentage;
-      } else {
-        throw new Error(
-          `The firstVisitPercentage option must be a number between 0 and 1. You passed in ${options.firstVisitPercentage}.`
-        );
-      }
-    } else {
-      throw new Error(
-        `The firstVisitPercentage option must be a number. You passed in a ${typeof options.firstVisitPercentage}.`
-      );
-    }
-  }
-
-  if (options?.returnVisitPercentage) {
-    if (typeof options.returnVisitPercentage === "number") {
-      if (
-        options.returnVisitPercentage > 0 &&
-        options.returnVisitPercentage < 1
-      ) {
-        adjustments.returnVisitPercentage = options.returnVisitPercentage;
-      } else {
-        throw new Error(
-          `The returnVisitPercentage option must be a number between 0 and 1. You passed in ${options.returnVisitPercentage}.`
-        );
-      }
-    } else {
-      throw new Error(
-        `The returnVisitPercentage option must be a number. You passed in a ${typeof options.returnVisitPercentage}.`
-      );
-    }
-  }
-
-  return adjustments;
-}
+import { parseOptions } from "./helpers/index.js";
 
 class CO2 {
   constructor(options) {
